rm trans_en

+++ src/db/scripts/backup_db_data.ts
@@ -1,7 +1,6 @@
 import * as fs from 'fs';
 import { z } from 'zod';
 import { dbClient_ext as db, queryClient } from '~/db/scripts/client';
-import { encrypt_text } from '~/tools/encrypt_decrypt';
 import { execSync } from 'child_process';
 
 const OUT_FOLDER = './backup';
@@ -27,30 +26,32 @@ async function main() {
   const JSON_TRANS_DATA_FILE = `${OUT_FOLDER}/translations.json`;
   fs.writeFileSync(JSON_TRANS_DATA_FILE, JSON.stringify(trans_data, null, 2));
 
-  // Encrypting User Data
-  const user_data = {
-    user_verification_requests: await db.query.user_verification_requests.findMany(),
-    users: await db.query.users.findMany()
-  };
-  const JSON_USER_DATA_FILE = `${OUT_FOLDER}/user_data_json_encrypted`;
-  fs.writeFileSync(
-    JSON_USER_DATA_FILE,
-    await encrypt_text(JSON.stringify(user_data, null, 2), envs.BACKUP_ENCRYPTION_KEY)
-  );
-
-  // Backup using pg_dump
-  execSync(
-    `pg_dump --dbname=${envs.PG_DATABASE_URL} --if-exists --clean --insert --no-owner --rows-per-insert=8000 -f b.sql`
-  );
-  const backup_file_data = fs.readFileSync('b.sql').toString('utf-8');
-  fs.writeFileSync(
-    `${OUT_FOLDER}/db_dump_sql_encrypted`,
-    await encrypt_text(backup_file_data, envs.BACKUP_ENCRYPTION_KEY),
-    {
+  if (!process.argv.slice(2).includes('--only-trans')) {
+    // Encrypting User Data
+    const user_data = {
+      user_verification_requests: await db.query.user_verification_requests.findMany(),
+      users: await db.query.users.findMany()
+    };
+    const JSON_USER_DATA_FILE = `${OUT_FOLDER}/user_data.json`;
+    fs.writeFileSync(JSON_USER_DATA_FILE, JSON.stringify(user_data, null, 2), {
       encoding: 'utf-8'
-    }
-  );
-  fs.rmSync('b.sql');
+    });
+
+    // Backup using pg_dump
+    execSync(
+      `pg_dump --dbname=${envs.PG_DATABASE_URL} --if-exists --clean --insert --no-owner --rows-per-insert=8000 -f b.sql`
+    );
+    const backup_file_data = fs.readFileSync('b.sql').toString('utf-8');
+    fs.writeFileSync(`${OUT_FOLDER}/db_dump.sql`, backup_file_data, {
+      encoding: 'utf-8'
+    });
+    fs.rmSync('b.sql');
+
+    // Zipping files
+    execSync(
+      `cd backup && zip -P ${envs.BACKUP_ENCRYPTION_KEY}  encrypted.zip db_dump.sql user_data.json && rm db_dump.sql user_data.json`
+    );
+  }
 }
 
 main().then(() => queryClient.end());
